relocate colorMap
