+++ src/modules/_shared/components/atoms/ColorSelectorButton/ColorSelectorButton.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import cn from 'classnames';
 import styles from './ColorSelector.module.scss';
 import { Color } from '../../../../../types/Color';
+import { ColorMapping } from '../../../utils/ColorMapping';
 
 interface Props {
   isActive?: boolean;
@@ -14,29 +15,6 @@ export const ColorSelectorButton: React.FC<Props> = ({
   color = 'red',
   click,
 }) => {
-  const colorMap: Record<string, string> = {
-    black: '#000',
-    gold: '#FFD700',
-    yellow: '#FFFF00',
-    green: '#008000',
-    midnightgreen: '#004953',
-    silver: '#c0c0c0',
-    spacegray: '#717378',
-    red: '#ff0000',
-    white: '#FFF',
-    purple: '#800080',
-    coral: '#ff7f50',
-    rosegold: '#B76E79',
-    midnight: '#191970',
-    spaceblack: '#505150',
-    blue: '#0000ff',
-    pink: '#ffc0cb',
-    sierrablue: '#BFDAF7',
-    graphite: '#41424C',
-    skyblue: '#87CEEB',
-    starlight: '#F8F9EC',
-  };
-
   return (
     <div
       className={cn(styles.colorSelector, {
@@ -45,8 +23,7 @@ export const ColorSelectorButton: React.FC<Props> = ({
     >
       <a
         style={{
-          backgroundColor:
-            colorMap[color.replaceAll(' ', '').replaceAll('-', '')],
+          backgroundColor: ColorMapping(color),
         }}
         className={styles.colorSelectorLink}
         onClick={click}
